Last tweaks to tests before moving async support
A couple more tests require changes to be independent from async loggers
and logger configurations.

The `CoreContextSelectors` class is not used, so we remove it.
