+++ log4j-core-test/src/main/java/org/apache/logging/log4j/core/async/QueueFullAbstractTest.java
@@ -19,17 +19,12 @@ package org.apache.logging.log4j.core.async;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.lmax.disruptor.RingBuffer;
-import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 import java.util.Stack;
 import java.util.concurrent.CountDownLatch;
-import java.util.stream.Collectors;
-import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.core.Appender;
 import org.apache.logging.log4j.core.LogEvent;
@@ -39,11 +34,8 @@ import org.apache.logging.log4j.core.config.Configuration;
 import org.apache.logging.log4j.core.config.LoggerConfig;
 import org.apache.logging.log4j.core.test.async.BlockingAppender;
 import org.apache.logging.log4j.core.util.Constants;
-import org.apache.logging.log4j.status.StatusData;
 import org.apache.logging.log4j.status.StatusLogger;
-import org.apache.logging.log4j.test.ListStatusListener;
 import org.apache.logging.log4j.test.junit.UsingStatusListener;
-import org.apache.logging.log4j.util.ReflectionUtil;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Timeout;
 import org.junit.jupiter.api.Timeout.ThreadMode;
@@ -88,35 +80,6 @@ public abstract class QueueFullAbstractTest {
         }
     }
 
-    protected static class DomainObject {
-
-        private final Logger innerLogger;
-        private final Unlocker unlocker;
-        private final int count;
-
-        public DomainObject(final Logger innerLogger, final Unlocker unlocker, final int loggingCount) {
-            this.innerLogger = innerLogger;
-            this.unlocker = unlocker;
-            this.count = loggingCount;
-        }
-
-        @Override
-        public String toString() {
-            for (int i = 0; i < count; i++) {
-                LOGGER.info(
-                        "DomainObject logging message {}. Ring buffer capacity was {}, countdown latch was {}.",
-                        i,
-                        asyncRemainingCapacity(innerLogger),
-                        unlocker.countDownLatch.getCount());
-                unlocker.countDownLatch.countDown();
-                innerLogger.info("Logging in toString() #{}", i);
-            }
-            return "Who's bad?!";
-        }
-    }
-
-    private ListStatusListener statusListener;
-
     protected void testNormalQueueFullKeepsMessagesInOrder(
             final LoggerContext ctx, final BlockingAppender blockingAppender) throws Exception {
         checkConfig(ctx);
@@ -129,10 +92,9 @@ public abstract class QueueFullAbstractTest {
         unlocker.join();
     }
 
-    protected void checkConfig(final LoggerContext ctx) throws Exception {}
+    protected abstract void checkConfig(final LoggerContext ctx) throws Exception;
 
-    protected static void asyncTest(
-            final Logger logger, final Unlocker unlocker, final BlockingAppender blockingAppender) {
+    protected void asyncTest(final Logger logger, final Unlocker unlocker, final BlockingAppender blockingAppender) {
         for (int i = 0; i < MESSAGE_COUNT; i++) {
             LOGGER.info(
                     "Test logging message {}. Ring buffer capacity was {}, countdown latch was {}.",
@@ -161,63 +123,6 @@ public abstract class QueueFullAbstractTest {
         assertThat(actual).isEmpty();
     }
 
-    public void testLoggingFromToStringCausesOutOfOrderMessages(
-            final LoggerContext ctx, final BlockingAppender blockingAppender) throws Exception {
-        checkConfig(ctx);
-        // Non-reusable messages will call `toString()` on the main thread and block it.
-        final Logger logger = ctx.getLogger(this.getClass());
-
-        blockingAppender.countDownLatch = new CountDownLatch(1);
-        final Unlocker unlocker = new Unlocker(new CountDownLatch(MESSAGE_COUNT - 1), blockingAppender);
-        unlocker.start();
-        asyncRecursiveTest(logger, unlocker, blockingAppender);
-        unlocker.join();
-    }
-
-    void asyncRecursiveTest(final Logger logger, final Unlocker unlocker, final BlockingAppender blockingAppender) {
-        for (int i = 0; i < 1; i++) {
-            LOGGER.info(
-                    "Test logging message {}. Ring buffer capacity was {}, countdown latch was {}.",
-                    i,
-                    asyncRemainingCapacity(logger),
-                    unlocker.countDownLatch.getCount());
-            unlocker.countDownLatch.countDown();
-            final DomainObject obj = new DomainObject(logger, unlocker, MESSAGE_COUNT - 1);
-            logger.info("Logging naughty object #{}: {}", i, obj);
-        }
-
-        LOGGER.info(
-                "Waiting for message delivery: blockingAppender.logEvents.count={}.",
-                blockingAppender.logEvents.size());
-        while (blockingAppender.logEvents.size() < MESSAGE_COUNT) {
-            Thread.yield();
-        }
-        LOGGER.info(
-                "All {} messages have been delivered: blockingAppender.logEvents.count={}.",
-                MESSAGE_COUNT,
-                blockingAppender.logEvents.size());
-
-        final StatusData mostRecentStatusData = statusListener
-                .findStatusData(Level.WARN)
-                .reduce((ignored, data) -> data)
-                .orElse(null);
-        assertThat(mostRecentStatusData).isNotNull();
-        assertThat(mostRecentStatusData.getLevel()).isEqualTo(Level.WARN);
-        assertThat(mostRecentStatusData.getFormattedStatus())
-                .contains("Log4j2 logged an event out of order to prevent deadlock caused by domain "
-                        + "objects logging from their toString method when the async queue is full");
-
-        final List<String> actual = blockingAppender.logEvents.stream()
-                .map(e -> e.getMessage().getFormattedMessage())
-                .collect(Collectors.toList());
-        final String[] expected = new String[MESSAGE_COUNT];
-        for (int i = 0; i < MESSAGE_COUNT - 1; i++) {
-            expected[i] = "Logging in toString() #" + i;
-        }
-        expected[MESSAGE_COUNT - 1] = "Logging naughty object #0: Who's bad?!";
-        assertThat(actual).hasSize(MESSAGE_COUNT).contains(expected);
-    }
-
     static Stack<String> transform(final List<LogEvent> logEvents) {
         final List<String> filtered = getMessages(logEvents);
         Collections.reverse(filtered);
@@ -234,71 +139,22 @@ public abstract class QueueFullAbstractTest {
         return filtered;
     }
 
-    static long asyncRemainingCapacity(final Logger logger) {
-        if (logger instanceof AsyncLogger asyncLogger) {
-            return Optional.ofNullable(asyncLogger.getAsyncLoggerDisruptor())
-                    .map(AsyncLoggerDisruptor::getRingBuffer)
-                    .map(RingBuffer::remainingCapacity)
-                    .orElse(0L);
-        } else {
-            final LoggerConfig loggerConfig = ((org.apache.logging.log4j.core.Logger) logger).get();
-            if (loggerConfig instanceof AsyncLoggerConfig asyncLoggerConfig) {
-                return Optional.ofNullable(
-                                (AsyncLoggerConfigDisruptor) asyncLoggerConfig.getAsyncLoggerConfigDelegate())
-                        .map(AsyncLoggerConfigDisruptor::getRingBuffer)
-                        .map(RingBuffer::remainingCapacity)
-                        .orElse(0L);
-            } else {
-                final Appender async = loggerConfig.getAppenders().get("async");
-                if (async instanceof AsyncAppender) {
-                    return ((AsyncAppender) async).getQueueCapacity();
-                }
-            }
+    protected long asyncRemainingCapacity(final Logger logger) {
+        final LoggerConfig loggerConfig = ((org.apache.logging.log4j.core.Logger) logger).get();
+        final Appender async = loggerConfig.getAppenders().get("async");
+        if (async instanceof AsyncAppender) {
+            return ((AsyncAppender) async).getQueueCapacity();
         }
         throw new IllegalStateException("Neither Async Loggers nor AsyncAppender are configured");
     }
 
-    protected static Field field(final Class<?> c, final String name) throws NoSuchFieldException {
-        final Field f = c.getDeclaredField(name);
-        ReflectionUtil.makeAccessible(f);
-        return f;
-    }
-
     protected static void assertAsyncAppender(final LoggerContext ctx) {
-        assertThat(ctx).isNotInstanceOf(AsyncLoggerContext.class);
-
         final Configuration config = ctx.getConfiguration();
         assertThat(config).isNotNull();
-        assertThat(config.getRootLogger()).isNotInstanceOf(AsyncLoggerConfig.class);
 
         final Collection<Appender> appenders =
                 config.getRootLogger().getAppenders().values();
         assertThat(appenders).hasSize(1).allMatch(AsyncAppender.class::isInstance);
-    }
-
-    protected static void assertAsyncLogger(final LoggerContext ctx, final int expectedBufferSize) {
-        assertThat(ctx).isInstanceOf(AsyncLoggerContext.class);
-        assertThat(((AsyncLoggerContext) ctx)
-                        .getAsyncLoggerDisruptor()
-                        .getRingBuffer()
-                        .getBufferSize())
-                .isEqualTo(expectedBufferSize);
-
-        final Configuration config = ctx.getConfiguration();
-        assertThat(config).isNotNull();
-        assertThat(config.getRootLogger()).isNotInstanceOf(AsyncLoggerConfig.class);
-    }
-
-    protected static void assertAsyncLoggerConfig(final LoggerContext ctx, final int expectedBufferSize) {
-        assertThat(ctx).isNotInstanceOf(AsyncLoggerContext.class);
-
-        final Configuration config = ctx.getConfiguration();
-        assertThat(config).isNotNull();
-        assertThat(config.getRootLogger()).isInstanceOf(AsyncLoggerConfig.class);
-        final DisruptorConfiguration disruptorConfig = config.getExtension(DisruptorConfiguration.class);
-        final AsyncLoggerConfigDisruptor disruptor =
-                (AsyncLoggerConfigDisruptor) disruptorConfig.getAsyncLoggerConfigDelegate();
-        assertThat(disruptor.getRingBuffer().getBufferSize()).isEqualTo(expectedBufferSize);
     }
 
     protected static void assertFormatMessagesInBackground() {
